+++ CotEditor/Sources/Settings Window/Panes/AppearanceSettingsView.swift
@@ -189,6 +189,10 @@ private struct FontSettingView: View {
                     .help(self.displayFontName)
                     .frame(maxWidth: 260)
                     .alignmentGuide(.firstTextBaseline) { $0.height }
+                FontSizeStepper(String(localized: "Font size", table: "AppearanceSettings"), font: self.font)
+                    .accessibilityValue(String(localized: "\(self.font.wrappedValue.pointSize, format: .number) points",
+                                               table: "AppearanceSettings", comment: "accessibility label for font size"))
+                    .padding(.leading, -4)
                 FontPicker(String(localized: "Select…", table: "AppearanceSettings", comment: "label for font picker button"), selection: self.font)
             }
             HStack {